Add multi-workplace shift assignment and employee management
- Support 5 workplaces (House 1-4, Sorting) with AM/PM input
- Handle optional inputs (remove required, add empty task option)
- Add PartTimeEmployee entity, controller, and UI (available_employees,
employee_form)
- Update logging with logback and build configurations
- Include entities (PreferredShift, ShiftAssignment, Task, Workplace)
and repositories

+++ src/main/java/com/example/app/controller/PartTimeEmployeeController.java
@@ -41,7 +41,7 @@ import com.example.app.service.PartTimeEmployeeService;
       }
 
       @PostMapping("/edit/{id}")
-      public String update(@PathVariable Integer id, PartTimeEmployee employee) {
+      public String update(@PathVariable Long id, PartTimeEmployee employee) {
           employee.setId(id);
           service.save(employee);
           return "redirect:/admin/employees";
